Replace WMS layers used in development with MVT layers
The WMS layers were being served by GeoServer, but that's a heavy
dependency with a lot of features that aren't necessary for our
purposes.

An MVT (Mapbox Vector Tile) view was added to the API that serves street
and intersection tiles directly from the database.

In dev/debug, the street and intersection MVT layers (as a layer group)
are set as the initial visible base layer.

In addition, in dev/debug, the OSM layer is uses the regular OSM source
instead of tile-debugging source. That wasn't very useful and the MVT
layers serve the purpose of not hitting the OSM tile layers (unless
necessary) in dev/debug.